+++ src/main.rs
@@ -31,9 +31,11 @@ use crate::cache::add_pool;
 use crate::config::CONFIG;
 use crate::routes::routes;
 use crate::sync::sync::Sync;
-use actix_web::{middleware, App, HttpServer};
+use actix_cors::Cors;
+use actix_web::{http, middleware, App, HttpServer};
 use async_std::task;
 use log::info;
+use std::env;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -57,9 +59,21 @@ async fn main() -> std::io::Result<()> {
     // Start http server
     let addr = format!("{}:{}", config.turboflakes_host, config.turboflakes_port);
     HttpServer::new(move || {
+        let cors = Cors::default()
+            .allowed_origin_fn(|origin, _req_head| {
+                let allowed_origin =
+                    env::var("TURBOFLAKES_CORS_ALLOW_ORIGIN").unwrap_or("localhost".to_string());
+                origin.as_bytes().ends_with(allowed_origin.as_bytes())
+            })
+            .allowed_methods(vec!["GET"])
+            .allowed_headers(vec![http::header::CONTENT_TYPE])
+            .supports_credentials()
+            .max_age(3600);
+
         App::new()
-            .configure(add_pool)
             .wrap(middleware::Logger::default())
+            .wrap(cors)
+            .configure(add_pool)
             .configure(routes)
     })
     .bind(addr)?