+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix inclusion rate and average reward points for a validator
+- Setup Cors and configurable allowed origin header
 
 ## [0.1.0] - 2021-04-19